#7 Merge finder.py
